+++ packages/frontend/src/components/DataViz/transformers/useTableDataModel.ts
@@ -50,11 +50,11 @@ export const useTableDataModel = <T extends ResultsRunner>({
             // react table has a bug with accessors that has dots in them
             // we found the fix here -> https://github.com/TanStack/table/issues/1671
             // do not remove the line below
-            accessorFn: (row) => row[column],
+            accessorFn: resultsRunner.getColumnsAccessorFn(column),
             header: config?.columns[column].label || column,
             cell: getValueCell,
         }));
-    }, [columns, config]);
+    }, [columns, config?.columns, resultsRunner]);
 
     const table = useReactTable({
         data: rows,